pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "docvec"
version = "0.1.0"
description = "Extract, chunk, and embed PDF documents and queries for use in semantic search"
requires-python = ">=3.11"
dependencies = [
    "pymupdf4llm==1.27.2.2",
    "sentence-transformers==5.3.0",
    "boto3==1.42.86",
    "python-dotenv==1.2.2",
    "langchain-text-splitters>=0.3",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0",
    "instructor>=1.0",
]
dev = [
    "pytest",
    "boto3-stubs[s3]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-s"
